Added my notebooks for GJ 1214 b !
This contains the spectra for GJ 1214 b, which are saved not as a time series but as a function of impact parameter b. They then need to be converted from b into time, and then the analysis from the hands-on sessions is applied to get the spectrum of GJ 1214 b.